requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
Brotli==1.1.0
celery==5.3.5
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cssselect2==0.7.0
Django==4.2.7
djangorestframework==3.14.0
fonttools==4.44.3
html5lib==1.1
idna==3.4
kombu==5.3.4
pdfkit==1.0.0
Pillow==10.1.0
prompt-toolkit==3.0.41
pycparser==2.21
pydyf==0.8.0
pyphen==0.14.0
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tinycss2==1.2.1
tzdata==2023.3
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.10
weasyprint==60.1
webencodings==0.5.1
wkhtmltopdf==0.2
zopfli==0.2.3
